Move TestConfig to testutils.
